Set loadFeed as nested function called twice in beforeEach
As the loadFeed functions are asynchronous, we need to make sure
they run one after another and not at the same time. So, we can
nest the second loadFeed call inside the first, using done() to
signal the end of the asynchronous work.

Instead of storing feed entry nodes into an array and comparing
each item in first feed versus each item in the next feed, we can
simply compare the innerHTML of the two feeds.

Also removed the done() function within the test as no
async code is running at this point.